+++ src/main/java/org/enviro/assessment/grad001/karabokhunou/investments/dto/CreditDebitRequest.java
@@ -12,5 +12,5 @@ import java.math.BigInteger;
 @AllArgsConstructor
 public class CreditDebitRequest {
     private  String accountNumber;
-    private BigInteger amount;
+    private BigDecimal amount;
 }
